moving home_page_test into home_page; introduce master merge github action; cleanup
