+++ .github/workflows/build-and-test.yml
@@ -1,0 +1,26 @@
+name: Build and Test
+on:
+  pull_request:
+    branches:
+      - master
+jobs:
+  build-and-test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v1
+      - name: Setup Java
+        uses: actions/setup-java@v1
+        with:
+          java-version: '12.x'
+      - name: Setup Flutter
+        uses: subosito/flutter-action@v1
+        with:
+          channel: 'stable'
+      - name: Pub Get Packages
+        run: flutter pub get
+      - name: Build AOT
+        run: flutter build aot
+      - name: Analyze Dart Code
+        run: flutter analyze
+      - name: Run Tests
+        run: flutter test
